Fixed imbalance calculation with period + bug in greedy optimization
fixed problem in _return_imb_with_coords - there you pass iteratively coords from a coord list. Lets say our original X data has 29 features, but as first coord in coordlist I passed only [1,3]. This is fed into the compute_nn_distances, together with self.period. self.period is an array with 29 numbers, while the coord is this example is only 2. This led finally to the problem in the cKDTree. Fixed.